backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.3
django-redis>=5.4
django-celery-results>=2.5
channels>=4.0
channels-redis>=4.2
psycopg[binary]>=3.1
celery>=5.3
redis>=5.0
qrcode>=7.4
Pillow>=10.0
python-dotenv>=1.0
whitenoise>=6.6
argon2-cffi>=23.1
hiredis>=2.3

dj-database-url>=2.1.0

drf-spectacular==0.27.2
drf-spectacular-sidecar
django-debug-toolbar>=4.4
